feat: 무한 스크롤에 로딩 구현 및 성능 개선
